Add crypto service
