+++ workers/loc.api/di/types.js
@@ -47,6 +47,7 @@ module.exports = {
   SubAccountApiData: Symbol.for('SubAccountApiData'),
   PositionsAudit: Symbol.for('PositionsAudit'),
   OrderTrades: Symbol.for('OrderTrades'),
+  Crypto: Symbol.for('Crypto'),
   Authenticator: Symbol.for('Authenticator'),
   PrivResponder: Symbol.for('PrivResponder')
 }
